requirements.txt
fastapi
sqlalchemy
datetime
pandas
bs4
uvicorn
requests
dotenv
jinja2
python-multipart
psycopg2-binary
python-telegram-bot